Using password files for OpenSSL
All helper scripts accepts now as argument a file which contains the
password to encrypt and decrypt the private key of CA.
The wrapping script build-ca.sh asks for needed passwords and stores
them in temp files. These files will be shreded after using.
This reduces the amount of passwords to be asked for.
Paranoid users can use a switch in the build-ca-sh script to avoid
saving the password in a temp file.